Fixed non-fatal typo in one of the battle() tests
For the test "ghoulGang just barely meets the threshold", the win threshold for the battle is 65. However, it should be 75 since each of the ghosts trick or treat for 2 candies and then rob each of the witches of 3 candies each (add 5 overall to each of their totals to bring it to 15). The test case still passes because 75 > 65, but it does not test the intended edge case. (Pointed out by Anonymous Cat on EdDiscussion)